backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
scikit-learn>=1.4.1
joblib==1.3.2
python-multipart==0.0.9
pydantic==2.6.4
python-dotenv==1.0.1
numpy==1.26.4
